backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib
pydantic
requests
